move retry function to utilities
